♻️ Legg til en TilsynBarnBeregningFellesService for felles funksjoner mellom V1 og V2
