+++ src/main/kotlin/no/nav/tilleggsstonader/sak/vedtak/barnetilsyn/beregning/TilsynBarnBeregningService.kt
@@ -2,15 +2,11 @@ package no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.beregning
 
 import no.nav.tilleggsstonader.sak.behandling.domain.Saksbehandling
 import no.nav.tilleggsstonader.sak.felles.domain.BarnId
-import no.nav.tilleggsstonader.sak.felles.domain.BehandlingId
-import no.nav.tilleggsstonader.sak.infrastruktur.database.repository.findByIdOrThrow
 import no.nav.tilleggsstonader.sak.infrastruktur.exception.feilHvis
-import no.nav.tilleggsstonader.sak.util.YEAR_MONTH_MIN
 import no.nav.tilleggsstonader.sak.util.datoEllerNesteMandagHvisLørdagEllerSøndag
-import no.nav.tilleggsstonader.sak.util.toYearMonth
 import no.nav.tilleggsstonader.sak.vedtak.TypeVedtak
-import no.nav.tilleggsstonader.sak.vedtak.VedtakRepository
 import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.beregning.TilsynBarnBeregningValideringUtil.validerPerioderForInnvilgelse
+import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.beregning.TilsynBeregningUtil.brukBeregningsgrunnlagFraOgMedRevurderFra
 import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.beregning.TilsynBeregningUtil.splitFraRevurderFra
 import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.beregning.TilsynBeregningUtil.tilAktiviteterPerMånedPerType
 import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.beregning.TilsynBeregningUtil.tilDagerPerUke
@@ -22,33 +18,23 @@ import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.domain.Beregningsgrunnlag
 import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.domain.BeregningsresultatForMåned
 import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.domain.BeregningsresultatTilsynBarn
 import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.domain.StønadsperiodeGrunnlag
-import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.domain.tilAktiviteter
 import no.nav.tilleggsstonader.sak.vedtak.domain.StønadsperiodeBeregningsgrunnlag
-import no.nav.tilleggsstonader.sak.vedtak.domain.VedtakTilsynBarn
-import no.nav.tilleggsstonader.sak.vedtak.domain.VedtakUtil.withTypeOrThrow
-import no.nav.tilleggsstonader.sak.vedtak.domain.beregningsresultat
 import no.nav.tilleggsstonader.sak.vedtak.domain.tilSortertStønadsperiodeBeregningsgrunnlag
 import no.nav.tilleggsstonader.sak.vilkår.stønadsperiode.domain.StønadsperiodeRepository
 import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.AktivitetType
-import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.ResultatVilkårperiode
-import no.nav.tilleggsstonader.sak.vilkår.vilkårperiode.domain.VilkårperiodeRepository
 import org.springframework.stereotype.Service
 import java.math.BigDecimal
-import java.math.RoundingMode
-import java.time.LocalDate
 
 /**
  * Stønaden dekker 64% av utgifterne til barnetilsyn
  */
 val DEKNINGSGRAD_TILSYN_BARN = BigDecimal("0.64")
-private val SNITT_ANTALL_VIRKEDAGER_PER_MÅNED = BigDecimal("21.67")
 
 @Service
 class TilsynBarnBeregningService(
     private val stønadsperiodeRepository: StønadsperiodeRepository,
-    private val vilkårperiodeRepository: VilkårperiodeRepository,
     private val tilsynBarnUtgiftService: TilsynBarnUtgiftService,
-    private val repository: VedtakRepository,
+    private val tilsynBarnBeregningFellesService: TilsynBarnBeregningFellesService,
 ) {
     fun beregn(
         behandling: Saksbehandling,
@@ -58,7 +44,8 @@ class TilsynBarnBeregningService(
             "Skal ikke beregne for avslag"
         }
         val perioder = beregnAktuellePerioder(behandling, typeVedtak)
-        val relevantePerioderFraForrigeVedtak = finnRelevantePerioderFraForrigeVedtak(behandling)
+        val relevantePerioderFraForrigeVedtak =
+            tilsynBarnBeregningFellesService.finnRelevantePerioderFraForrigeVedtak(behandling)
         return BeregningsresultatTilsynBarn(relevantePerioderFraForrigeVedtak + perioder)
     }
 
@@ -77,33 +64,19 @@ class TilsynBarnBeregningService(
                 .tilSortertStønadsperiodeBeregningsgrunnlag()
                 .splitFraRevurderFra(behandling.revurderFra)
 
-        val aktiviteter = finnAktiviteter(behandling.id)
+        val aktiviteter = tilsynBarnBeregningFellesService.finnAktiviteter(behandling.id)
 
         validerPerioderForInnvilgelse(stønadsperioder, aktiviteter, utgifterPerBarn, typeVedtak, behandling.revurderFra)
 
         val beregningsgrunnlag =
             lagBeregningsgrunnlagPerMåned(stønadsperioder, aktiviteter, utgifterPerBarn)
-                .brukPerioderFraOgMedRevurderFra(behandling.revurderFra)
+                .brukBeregningsgrunnlagFraOgMedRevurderFra(behandling.revurderFra)
         return beregn(beregningsgrunnlag)
     }
 
-    private fun finnRelevantePerioderFraForrigeVedtak(behandling: Saksbehandling): List<BeregningsresultatForMåned> =
-        behandling.forrigeBehandlingId?.let { forrigeBehandlingId ->
-            val beregningsresultat =
-                repository
-                    .findByIdOrThrow(forrigeBehandlingId)
-                    .withTypeOrThrow<VedtakTilsynBarn>()
-                    .data
-                    .beregningsresultat()
-                    ?: error("Finner ikke beregningsresultat på vedtak for behandling=$forrigeBehandlingId")
-            val revurderFraMåned = behandling.revurderFra?.toYearMonth() ?: YEAR_MONTH_MIN
-
-            beregningsresultat.perioder.filter { it.grunnlag.måned < revurderFraMåned }
-        } ?: emptyList()
-
     private fun beregn(beregningsgrunnlag: List<Beregningsgrunnlag>): List<BeregningsresultatForMåned> =
         beregningsgrunnlag.map {
-            val dagsats = beregnDagsats(it)
+            val dagsats = tilsynBarnBeregningFellesService.beregnDagsats(it)
             val beløpsperioder = lagBeløpsperioder(dagsats, it)
 
             BeregningsresultatForMåned(
@@ -123,36 +96,10 @@ class TilsynBarnBeregningService(
             val dato = it.stønadsperiode.fom.datoEllerNesteMandagHvisLørdagEllerSøndag()
             Beløpsperiode(
                 dato = dato,
-                beløp = beregnBeløp(dagsats, it.antallDager),
+                beløp = tilsynBarnBeregningFellesService.beregnBeløp(dagsats, it.antallDager),
                 målgruppe = it.stønadsperiode.målgruppe,
             )
         }
-
-    private fun beregnBeløp(
-        dagsats: BigDecimal,
-        antallDager: Int,
-    ) = dagsats
-        .multiply(antallDager.toBigDecimal())
-        .setScale(0, RoundingMode.HALF_UP)
-        .toInt()
-
-    /**
-     * Divide trenger en scale som gir antall desimaler på resultatet fra divideringen
-     * Sånn sett blir `setScale(2, RoundingMode.HALF_UP)` etteråt unødvendig
-     * Tar likevel med den for å gjøre det tydelig at resultatet skal maks ha 2 desimaler
-     */
-    private fun beregnDagsats(grunnlag: Beregningsgrunnlag): BigDecimal {
-        val utgifter = grunnlag.utgifterTotal.toBigDecimal()
-        val utgifterSomDekkes =
-            utgifter
-                .multiply(DEKNINGSGRAD_TILSYN_BARN)
-                .setScale(0, RoundingMode.HALF_UP)
-                .toInt()
-        val satsjusterteUtgifter = minOf(utgifterSomDekkes, grunnlag.makssats).toBigDecimal()
-        return satsjusterteUtgifter
-            .divide(SNITT_ANTALL_VIRKEDAGER_PER_MÅNED, 2, RoundingMode.HALF_UP)
-            .setScale(2, RoundingMode.HALF_UP)
-    }
 
     private fun lagBeregningsgrunnlagPerMåned(
         stønadsperioder: List<StønadsperiodeBeregningsgrunnlag>,
@@ -225,24 +172,4 @@ class TilsynBarnBeregningService(
                 BeregningsgrunnlagUtils.beregnAntallAktivitetsdagerForUke(periode, aktiviteterForUke)
             }.sum()
     }
-
-    /**
-     * Dersom man har satt revurderFra så skal man kun beregne perioder fra og med den måneden
-     * Hvis vi eks innvilget 1000kr for 1-31 august, så mappes hele beløpet til 1 august.
-     * Dvs det lages en andel som har fom-tom 1-1 aug
-     * Når man revurderer fra midten på måneden og eks skal endre målgruppe eller aktivitetsdager,
-     * så har man allerede utbetalt 500kr for 1-14 august, men hele beløpet er ført på 1 aug.
-     * For at beregningen då skal bli riktig må man ha med grunnlaget til hele måneden og beregne det på nytt, sånn at man får en ny periode som er
-     * 1-14 aug, 500kr, 15-30 aug 700kr.
-     */
-    private fun List<Beregningsgrunnlag>.brukPerioderFraOgMedRevurderFra(revurderFra: LocalDate?): List<Beregningsgrunnlag> {
-        val revurderFraMåned = revurderFra?.toYearMonth() ?: return this
-
-        return this.filter { it.måned >= revurderFraMåned }
-    }
-
-    private fun finnAktiviteter(behandlingId: BehandlingId): List<Aktivitet> =
-        vilkårperiodeRepository
-            .findByBehandlingIdAndResultat(behandlingId, ResultatVilkårperiode.OPPFYLT)
-            .tilAktiviteter()
 }
