+++ src/test/kotlin/no/nav/tilleggsstonader/sak/vedtak/barnetilsyn/TilsynBarnBeregnYtelseStegTest.kt
@@ -17,6 +17,7 @@ import no.nav.tilleggsstonader.sak.vedtak.OpphørValideringService
 import no.nav.tilleggsstonader.sak.vedtak.VedtakRepository
 import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.TilsynBarnTestUtil.innvilgelseDto
 import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.TilsynBarnTestUtil.opphørDto
+import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.beregning.TilsynBarnBeregningFellesService
 import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.beregning.TilsynBarnBeregningService
 import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.beregning.TilsynBarnUtgiftService
 import no.nav.tilleggsstonader.sak.vedtak.barnetilsyn.beregning.UtgiftBeregning
@@ -41,12 +42,17 @@ class TilsynBarnBeregnYtelseStegTest {
     private val tilsynBarnUtgiftService = mockk<TilsynBarnUtgiftService>(relaxed = true)
     private val opphørValideringService = mockk<OpphørValideringService>(relaxed = true)
 
+    val tilsynBarnBeregningFellesService =
+        TilsynBarnBeregningFellesService(
+            vilkårperiodeRepository = vilkårperiodeRepository,
+            vedtakRepository = repository,
+        )
+
     val tilsynBarnBeregningService =
         TilsynBarnBeregningService(
             stønadsperiodeRepository = stønadsperiodeService,
-            vilkårperiodeRepository = vilkårperiodeRepository,
             tilsynBarnUtgiftService = tilsynBarnUtgiftService,
-            repository = repository,
+            tilsynBarnBeregningFellesService = tilsynBarnBeregningFellesService,
         )
     val steg =
         TilsynBarnBeregnYtelseSteg(
